refactor: improve test structure and use tags for choosing tests
- Split test helpers into separate files for better organization:
  - helper_mock_test.go: Contains mock server implementation for unit tests
  - helper_integration_test.go: Contains setup for integration tests
- Add build tags to separate unit and integration tests
- Add helper functions for test environment configuration
