+++ download_history/download_history_test.go
@@ -1,6 +1,3 @@
-//go:build !integration
-// +build !integration
-
 package download_history
 
 import (
